Extract the non-POST blocking logic
I've tried to write this as well as I can, but it seems unavoidably
awkward to have an HTTP handler that _might_ respond.

However, I have to do this if I want a generic limit wrapper that can
cover multiple RPC handlers on a single mux while maintaining the same
output format. This check is important enough that I'd rather duplicate
it than require this limiter to always exist.